+++ src/Vehicle.hpp
@@ -17,14 +17,14 @@ private:
   static std::vector<std::shared_ptr<VehicleType>> _vehicleType;
   uint8_t _index;
   uint16_t _position;
-  int16_t _previousPosition = -1;
+  int _previousPosition = -1;
   int _street = -1;
   int _timePenalty = 0;
   double _velocity = 0.;
   int _timeTraveled = 0;
 
 public:
-  Vehicle(uint8_t); //!< create a vehicle of a type in _vehicleType
+  Vehicle(int); //!< create a vehicle of a type in _vehicleType
   Vehicle(const Vehicle &v)
       : _index(v._index), _position(v._position),
         _previousPosition(v._previousPosition), _street(v._street),
@@ -32,8 +32,8 @@ public:
         _timeTraveled(v._timeTraveled) {}
   ~Vehicle() = default;
 
-  static void addVehicleType(uint16_t,
-                             uint16_t); //!< add a vehicle type in _vehicleType
+  static void addVehicleType(int,
+                             int); //!< add a vehicle type in _vehicleType
   static void addVehicleType(
       std::string); //!< add many vehicle types in _vehicleType from file
   static std::shared_ptr<VehicleType>
@@ -42,9 +42,9 @@ public:
   getNVehicleType(); //!< get the number of vehicle types in _vehicleType
   uint8_t getType()
       const noexcept; // return the index of the vehicle type in _vehicleType
-  void setPosition(int16_t);
+  void setPosition(int);
   uint16_t getPosition() const noexcept;
-  int16_t getPreviousPosition() const noexcept;
+  int getPreviousPosition() const noexcept;
   void setStreet(int);
   int getStreet() const;
   uint16_t getDestination() const noexcept;
